C++17

+++ src/resonanceReconstruction/breitWigner/multilevel/Resonance/test/Resonance.test.cpp
@@ -102,7 +102,8 @@ SCENARIO("a MLBW resonance reconstruction"){
                        + fwhmFission;
 
       THEN("the shape will be correct"){
-        REQUIRE( 0.5 * resonanceTotal == Approx( fwhmTotal ) );
+        REQUIRE( static_cast< double >( 0.5 * resonanceTotal ) == 
+                 Approx( static_cast< double >( fwhmTotal ) ) );
       }
       
       THEN("the magnitude will match an alternative formulation"){
@@ -128,7 +129,8 @@ SCENARIO("a MLBW resonance reconstruction"){
                    + neutronWidth, Ratio<2> )
               + 4. * pow( energy - primedResonanceEnergy, Ratio<2> ) );
 
-        REQUIRE( reference == Approx(fwhmCapture) );
+        REQUIRE( static_cast< double >( reference ) == 
+                 Approx( static_cast< double >( fwhmCapture ) ) );
       }
     }
     
@@ -143,7 +145,8 @@ SCENARIO("a MLBW resonance reconstruction"){
                        + fwhmFission;
 
       THEN("the shape will be correct"){
-        REQUIRE( 0.5 * resonanceTotal == Approx( fwhmTotal ) );
+        REQUIRE( static_cast< double >( 0.5 * resonanceTotal ) == 
+                 Approx( static_cast< double >( fwhmTotal ) ) );
       }
       
       THEN("the magnitude will match an alternative formulation"){
@@ -169,7 +172,8 @@ SCENARIO("a MLBW resonance reconstruction"){
                    + neutronWidth, Ratio<2> )
               + 4. * pow( energy - primedResonanceEnergy, Ratio<2> ) );
 
-        REQUIRE( reference == Approx(fwhmCapture) );
+        REQUIRE( static_cast< double >( reference ) == 
+                 Approx( static_cast< double >( fwhmCapture ) ) );
       }
     }    
   }
